+++ src/main/java/com/enotes/controller/UserController.java
@@ -5,11 +5,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.enotes.dto.PasswordChngRequest;
 import com.enotes.dto.UserResponse;
 import com.enotes.entity.User;
+import com.enotes.service.AuthService;
+import com.enotes.service.UserService;
 import com.enotes.util.commonUtil;
 
 @RestController
@@ -19,6 +24,10 @@ public class UserController {
 	@Autowired
 	private ModelMapper mapper;
 	
+	@Autowired
+	private UserService userService;
+
+	
 	@GetMapping("/profile")
 	public ResponseEntity<?> getProfile()
 	{
@@ -26,4 +35,11 @@ public class UserController {
 		UserResponse userResponse = mapper.map(loggedInUser, UserResponse.class);
 		return commonUtil.createBuildResponse(userResponse, HttpStatus.OK);
 	}
+	
+	@PostMapping("/change-password")
+	public ResponseEntity<?> changePassword(@RequestBody PasswordChngRequest passwordRequest) {
+		userService.changePassword(passwordRequest);
+		return commonUtil.createBuildResponseMessage("Password change success", HttpStatus.OK);
+	}
+
 }
