+++ src/main/java/com/enotes/dto/LoginResponse.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class LoginResponse {
-	private UserDto user;
+	private UserResponse user;
 
 	private String token;
