+++ src/main/java/com/enotes/dto/PasswordChngRequest.java
@@ -1,0 +1,14 @@
+package com.enotes.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PasswordChngRequest {
+private String oldPassword;
+	
+	private String newPassword;
+}
