+++ src/main/java/com/enotes/dto/UserRequest.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class UserDto {
+public class UserRequest {
 
 	private Integer id;
