+++ src/main/java/com/enotes/service/AuthService.java
@@ -1,0 +1,12 @@
+package com.enotes.service;
+
+import com.enotes.dto.LoginRequest;
+import com.enotes.dto.LoginResponse;
+import com.enotes.dto.UserRequest;
+
+public interface AuthService {
+
+	Boolean register(UserRequest userRequest, String url) throws Exception;
+
+	LoginResponse login(LoginRequest loginRequest);
+}
