+++ src/main/java/com/enotes/service/UserService.java
@@ -1,12 +1,7 @@
 package com.enotes.service;
 
-import com.enotes.dto.LoginRequest;
-import com.enotes.dto.LoginResponse;
-import com.enotes.dto.UserDto;
+import com.enotes.dto.PasswordChngRequest;
 
 public interface UserService {
-
-	Boolean register(UserDto userDto, String url) throws Exception;
-
-	LoginResponse login(LoginRequest loginRequest);
+	public void changePassword(PasswordChngRequest passwordRequest);
 }
